refactor: extract components from SettingsAuthPage into own directory
