+++ frontend/src/features/admin-form/settings/components/AuthSettingsSection/index.ts
@@ -1,0 +1,1 @@
+export { AuthSettingsSectionContainer as default } from './AuthSettingsSectionContainer'
